Add files via upload
I am what you call a perfectionist. I always strive for perfection, and I appreciate everyone and everything that is perfect. That is why I have recently acquired an appreciation for perfect numbers! I absolutely need to know which numbers from 1 to 1000 are considered perfect. From what I recall, a perfect number is a positive integer that is equal to the sum of all its divisors other than itself.

Example:
6 is a perfect number because 6 = 1 + 2 + 3